requirements.txt
# PDF Processing
PyMuPDF==1.23.22
pytesseract==0.3.10
Pillow==10.2.0

# Data Validation and Models
pydantic==2.6.1
pydantic-settings==2.1.0

# API Framework
fastapi==0.109.2
uvicorn==0.27.1

# HTTP Client
httpx==0.26.0

# AI/LLM Integration
openai==1.12.0
anthropic==0.17.0

# Utilities
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1

# Testing
pytest>=7.0.0,<8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.0.0

# Code Quality
ruff==0.2.1
mypy==1.8.0
black==24.1.1

# Background Tasks
celery==5.3.6
redis==5.0.1
